Push predicate to package level, eliminate unnecessary function
